+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYKQ9874S3GQN77MXHD8P2J2"
+        "01KYM1Q85G01JE432X43MF2MQP"
     )
     r"""The unique identifier of the data source"""