+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-28T06:40:47.753Z"
+    created: Optional[str] = "2026-07-28T09:43:05.279Z"
 
-    updated: Optional[str] = "2026-07-28T06:40:47.753Z"
+    updated: Optional[str] = "2026-07-28T09:43:05.279Z"
 
     updated_by_id: OptionalNullable[str] = UNSET