+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYKQ986YJNA1EJFCSXKG5VQF"
+        "01KYM1Q85CZMVG7MS1G2067BRZ"
     )
     r"""The unique identifier of the data source"""