+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYKQ98722TGYNWQFY6TJQGPA"
+        "01KYM1Q85E5PFA2VV8W5JKXDEQ"
     )
     r"""The unique identifier of the data source"""