+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYKQ98788SPY61FS52WKJ4K6"
+        "01KYM1Q85MFVYXXYEVC0803VWM"
     )
     r"""The unique identifier of the data source"""